requirements.txt
flask
pymongo
requests
flask-swagger-ui
python-crontab